+++ lib/src/ats/ats.dart
@@ -130,3 +130,7 @@ part 'src/stock_closing/entity/stock_closing.dart';
 part 'src/stock_closing/entity/stock_closing_tank.dart';
 part 'src/stock_closing/input/stock_closing_input.dart';
 part 'src/stock_closing/input/stock_closing_tank_input.dart';
+
+// ATS Outbound Services
+part 'src/ats_outbound_services/ats_stream_model.dart';
+part 'src/ats_outbound_services/ats_stream_exit.dart';